requirements.txt
Django>=1.8,<1.9
wagtail==1.3.1
django-libsass==0.3
django-countries==3.4.1
django-model-utils==2.5
django-webpack-loader==0.3.0
django-filter==0.13.0
django-widget-tweaks==1.4.1
pytest-django==2.9.1
pytest-cov==2.2.1
pytest-factoryboy==1.1.6
psycopg2==2.6.1
gunicorn==19.6
whitenoise==3.2
django-autofixture==0.12.1
easy-thumbnails==2.3
django-parler==1.6.4